Refactor write_callback, change internal buffer to a stream-like callback, fix #25
Previously, instead of setting a python buffer to receive data by
curl's builtin write callback, we use a fixed callback function
and buffer to receive data, and after the full response content
is received, the data is copied back to user's buffer.

In the previous way, memory is wasted. More importantly, it's just
not possible to implement streaming.